feat: enhance typing experience
The very magic keymaps are conflict with the blink.cmp plugin, with
which the very magic may be not necessary any more. So just remove those
keymaps.

Display a color column with '+1' by default so that it would be easiser
to control the line width. Correspondingly add default text width for
C and C++.

Documentation comments are very important, so make it red.